unsafe handling of TZ environment variable
The TZ environment variable was passed through unchecked.  Most libc
tzset() implementations support passing an absolute pathname in the time
zone to point to an arbitrary, user-controlled file.  This may be used
to exploit bugs in the C library's TZ parser or open files the user
would not otherwise have access to.  Arbitrary file access via TZ could
also be used in a denial of service attack by reading from a file or
fifo that will block.